requirements.txt
# Python 3.6.8+ compatible versions - No Redis required
pytenable==1.3.4
python-dotenv==0.19.2
requests==2.27.1

